Refactor Extractor Tests
Split into multiple files based on test purpose
Use DynamicData to reduce duplication of test data rows.
Fix potential multi thread collision on resource governor.
